requirements.txt
--extra-index-url https://download.pytorch.org/whl/cu121  
torch==2.2.0 
torchvision==0.17.0 
torchaudio==2.2.0  

numpy==1.26.4 
Pillow==10.0.0 
pyyaml==6.0 
tqdm==4.66.0 
matplotlib==3.7.2 

tensorboard==2.14.0 
pytorch-lightning==2.2.1 

einops==0.7.0 
ema_pytorch==0.2.2 
accelerate==0.24.0